+++ test/transformer-attributify-jsx.test.ts
@@ -4,7 +4,6 @@ import presetWind3 from '@unocss/preset-wind3'
 import transformerAttributifyJsx from '@unocss/transformer-attributify-jsx'
 import MagicString from 'magic-string'
 import { describe, expect, it } from 'vitest'
-import { attributifyJsxOxcResolver } from '../packages-presets/transformer-attributify-jsx/src/resolver/oxc'
 
 const originalCode = `
 <div h-full text-center flex select-none className={red ? 'text-red': 'text-green'}>
@@ -282,24 +281,6 @@ describe('transformerAttributifyJsx', async () => {
                 </svg>
               )}
             />"
-    `)
-  })
-
-  it('oxc error', async () => {
-    const errorCode = '<d iv></div>'
-    const code = new MagicString(errorCode)
-    const transform = attributifyJsxOxcResolver({ code, id: 'app.tsx', uno: { uno, tokens: new Set() } as any, isBlocked: () => false })
-    await expect(transform).rejects.toThrowErrorMatchingInlineSnapshot(`
-      [Error: Oxc parse errors:
-
-        x Expected corresponding JSX closing tag for 'd'.
-         ,-[app.tsx:1:9]
-       1 | <d iv></div>
-         :  |      ^|^
-         :  |       \`-- Expected \`</d>\`
-         :  \`-- Opened here
-         \`----
-      ]
     `)
   })
 })
